+++ tests/telegram/test_admin.py
@@ -24,8 +24,8 @@ from src.credits import (
 )
 from src.dto import AlertState, MonthlyStats, UserCredits, UserTier, WitUsageStats
 from src.mongo import add_user_role, get_bot_config, get_users_by_role, remove_user_role
-from src.telegram.account_handlers import mystats_command
-from src.telegram.admin import (
+from src.telegram.handlers.account import mystats_command
+from src.telegram.handlers.admin import (
     _build_providers_panel,
     add_credits_command,
     add_tester_command,
@@ -37,7 +37,7 @@ from src.telegram.admin import (
     remove_vip_command,
     unblock_command,
 )
-from src.telegram.handlers import stats_command
+from src.telegram.handlers.common import stats_command
 
 
 class TestAdminRole: